backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3

SQLAlchemy==2.0.32
pydantic==2.11.3
pydantic-settings==2.3.4
email-validator==2.2.0

python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1

pymysql==1.1.1
httpx==0.27.0